# apispec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/projects/projects/create": {
      "post": {
        "tags": [
          "Project Management",
          "Project Management"
        ],
        "summary": "Create",
        "description": "Create a new project.",
        "operationId": "create_projects_projects_create_post",
        "parameters": [
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCreate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/projects/update": {
      "put": {
        "tags": [
          "Project Management",
          "Project Management"
        ],
        "summary": "Update",
        "description": "Update an existing project.",
        "operationId": "update_projects_projects_update_put",
        "parameters": [
          {
            "name": "pid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/projects/": {
      "get": {
        "tags": [
          "Project Management",
          "Project Management"
        ],
        "summary": "List",
        "description": "List all projects for a given user.",
        "operationId": "list_projects_projects__get",
        "parameters": [
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectInfo"
                  },
                  "title": "Response List Projects Projects  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Project Management",
          "Project Management"
        ],
        "summary": "Delete",
        "description": "Delete a project by ID.",
        "operationId": "delete_projects_projects__delete",
        "parameters": [
          {
            "name": "pid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/training/yolo/detection": {
      "post": {
        "tags": [
          "Training",
          "Training"
        ],
        "summary": "Yolo Training",
        "description": "Start YOLO training with the provided parameters.",
        "operationId": "yolo_training_training_yolo_detection_post",
        "parameters": [
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YoloDetTrainingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/training/yolo/result": {
      "post": {
        "tags": [
          "Training",
          "Training"
        ],
        "summary": "Yolo Training Result",
        "description": "Handle the result of YOLO training.",
        "operationId": "yolo_training_result_training_yolo_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YoloHandlingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/validation/yolo/{vid}": {
      "get": {
        "tags": [
          "Validation",
          "Validation"
        ],
        "summary": "Get Validation Info",
        "description": "Get validation status and results",
        "operationId": "get_validation_info_validation_yolo__vid__get",
        "parameters": [
          {
            "name": "vid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/validation/yolo/result": {
      "post": {
        "tags": [
          "Validation",
          "Validation"
        ],
        "summary": "Receive Result",
        "description": "Receive validation completion callback and persist to Mongo/MinIO.",
        "operationId": "receive_result_validation_yolo_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Result"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreate"
      },
      "ProjectInfo": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "last_trn_snap": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Trn Snap"
          },
          "last_opt_snap": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Opt Snap"
          },
          "last_val_snap": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Val Snap"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "_id",
          "uid",
          "pid",
          "name",
          "status",
          "created_at"
        ],
        "title": "ProjectInfo"
      },
      "ProjectUpdate": {
        "properties": {
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "pid"
        ],
        "title": "ProjectUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "YoloDetTrainingRequest": {
        "properties": {
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "cid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cid"
          },
          "did": {
            "type": "string",
            "title": "Did"
          },
          "origin_tid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Tid"
          },
          "parameters": {
            "$ref": "#/components/schemas/YoloDetTrainingParams"
          }
        },
        "type": "object",
        "required": [
          "pid",
          "did",
          "parameters"
        ],
        "title": "YoloDetTrainingRequest"
      },
...
      "YoloHandlingRequest": {
        "properties": {
          "workdir": {
            "type": "string",
            "title": "Workdir"
          },
          "result": {
            "$ref": "#/components/schemas/YoloTrainingResult"
          }
        },
        "type": "object",
        "required": [
          "workdir",
          "result"
        ],
        "title": "YoloHandlingRequest"
      },
      "YoloTrainingResult": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "origin_tid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Tid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "workdir": {
            "type": "string",
            "title": "Workdir"
          },
          "started_time": {
            "type": "string",
            "title": "Started Time"
          },
          "completed_time": {
            "type": "string",
            "title": "Completed Time"
          },
          "parameters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameters"
          },
          "did": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Did"
          },
          "cid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cid"
          },
          "artifacts_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts Path"
          },
          "error_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Details"
          }
        },
        "type": "object",
        "required": [
          "uid",
          "pid",
          "status",
          "task_type",
          "workdir",
          "started_time",
          "completed_time"
        ],
        "title": "YoloTrainingResult"
      }
    }
  }
}